+++ lib/bounding_box.rb
@@ -3,17 +3,17 @@ class BoundingBox
               :right, :top
 
   def initialize(left, bottom, width, height)
-    @left = left.to_i
-    @bottom = bottom.to_i
-    @width = width.to_i
-    @height = height.to_i
+    @left = left.to_f
+    @bottom = bottom.to_f
+    @width = width.to_f
+    @height = height.to_f
     @right = @left + @width
     @top = @bottom + @height
   end
 
   def contains_point?(x,y)
-    x = x.to_i
-    y = y.to_i
+    x = x.to_f
+    y = y.to_f
     if left <= x && x <= right && bottom <= y && y <= top
       true
     else
